refactor: 테스트 디렉토리 구조 변경
