+++ src/main/java/domain/player/Players.java
@@ -1,6 +1,5 @@
 package domain.player;
 
-import domain.MoveType;
 import domain.ladder.Line;
 
 import java.util.ArrayList;
