+++ src/test/java/domain/player/MoveTypeTest.java
@@ -1,8 +1,7 @@
-package domain;
+package domain.player;
 
 import domain.ladder.Line;
 import domain.booleangenerator.TestBooleanGenerator;
-import domain.player.Player;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
