+++ src/test/java/domain/player/PlayerTest.java
@@ -1,6 +1,5 @@
-package domain;
+package domain.player;
 
-import domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
